requirements.txt
mkdocs==1.6.1
mkdocs-material==9.5.21
pymdown-extensions==10.15
pandas==2.3.0
numpy==2.1.2
matplotlib==3.9.2
seaborn==0.13.2
plotly==6.2.0
japanize-matplotlib==1.1.3
PyYAML==6.0.2
typing_extensions==4.12.2
Jinja2==3.1.4
psutil==5.9.8